This PR changes the implementation of the resolver blacklist to be fully lock-free, using an immutable cache that gets rebuilt every time a new address gets blacklisted.

Here are numbers for benchmark comparing master and the changes in this branch:
```
goos: darwin
goarch: amd64
pkg: github.com/segmentio/consul-go
BenchmarkResolverBlacklist/empty            50000000        31.6 ns/op       0 B/op       0 allocs/op
BenchmarkResolverBlacklist/empty-8          20000000        83.9 ns/op       0 B/op       0 allocs/op
BenchmarkResolverBlacklist/empty-32         20000000        84.5 ns/op       0 B/op       0 allocs/op
BenchmarkResolverBlacklist/filled           10000000       155 ns/op       0 B/op       0 allocs/op
BenchmarkResolverBlacklist/filled-8         10000000       126 ns/op       0 B/op       0 allocs/op
BenchmarkResolverBlacklist/filled-32        10000000       123 ns/op       0 B/op       0 allocs/op
PASS
ok  github.com/segmentio/consul-go9.628s
```
```
goos: darwin
goarch: amd64
pkg: github.com/segmentio/consul-go
BenchmarkResolverBlacklist/empty            300000000         4.47 ns/op       0 B/op       0 allocs/op
BenchmarkResolverBlacklist/empty-8          1000000000         3.58 ns/op       0 B/op       0 allocs/op
BenchmarkResolverBlacklist/empty-32         1000000000         2.50 ns/op       0 B/op       0 allocs/op
BenchmarkResolverBlacklist/filled           10000000       138 ns/op       0 B/op       0 allocs/op
BenchmarkResolverBlacklist/filled-8         20000000        69.6 ns/op       0 B/op       0 allocs/op
BenchmarkResolverBlacklist/filled-32        20000000        70.4 ns/op       0 B/op       0 allocs/op
PASS
ok  github.com/segmentio/consul-go12.964s
```
```
benchmark                                old ns/op     new ns/op     delta
BenchmarkResolverBlacklist/empty         31.6          4.47          -85.85%
BenchmarkResolverBlacklist/empty-8       83.9          3.58          -95.73%
BenchmarkResolverBlacklist/empty-32      84.5          2.50          -97.04%
BenchmarkResolverBlacklist/filled        155           138           -10.97%
BenchmarkResolverBlacklist/filled-8      126           69.6          -44.76%
BenchmarkResolverBlacklist/filled-32     123           70.4          -42.76%

```
I made this change because this code is on a very hot path of services that use this package for service discovery and load balancing and they were spending a lot of time dealing with the RWMutex.

Please take a look and let me know if anything should be changed.